+ + Config metric as separated (configuration) internal package
